+++ src/wai/annotations/tf/utils/_polygon_from_mask.py
@@ -8,13 +8,17 @@ from wai.common.geometry import Polygon, Point
 from ...image_utils import mask_to_polygon
 
 
-def polygon_from_mask(mask: bytes, view: Tuple[float, float, float, float]) -> Polygon:
+def polygon_from_mask(mask: bytes,
+                      view: Tuple[float, float, float, float],
+                      sample_stride: int = 1) -> Polygon:
     """
     Converts the PNG format mask into a polygon.
 
-    :param mask:    The PNG mask.
-    :param view:    The area inside the mask to look for the polygon (left, top, right, bottom).
-    :return:        A polygon.
+    :param mask:            The PNG mask.
+    :param view:            The area inside the mask to look for the polygon (left, top, right, bottom).
+    :param sample_stride:   The stride to step through the mask when finding the polygon. Results
+                            in a more coarsely-grained polygon but finishes faster.
+    :return:                A polygon.
     """
     # Decode the PNG
     image = PIL.Image.open(io.BytesIO(mask))
@@ -25,6 +29,7 @@ def polygon_from_mask(mask: bytes, view: Tuple[float, float, float, float]) -> P
     # Use find_contours to find the (only) polygon in the bit-mask
     poly_array = mask_to_polygon(binary_mask, 0.9,
                                  view=view,
+                                 mask_nth=sample_stride,
                                  fully_connected="high")[0]
 
     # Convert to a Polygon